11.1 ch8 숫자 - 문자열을 숫자 타입으로 변환하기 - 8.3 price 인자를 Double 타입으로 변환하기

+++ NyetHack/src/main/kotlin/Tavern.kt
@@ -42,7 +42,7 @@ private fun placeOrder(menuData: String) {
     val message = "마드리갈은 금화 $price 로 $name ($type)를 구입한다."
     println(message)
 
-    performPurchase(price)
+    performPurchase(price.toDouble())
 
     val phrase = if (name == "Dragon's Breath") {
         "마드리갈이 감탄한다: ${toDragonSpeak("와, $name 진짜 좋구나!")}"
